pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "wekare-shared"
version = "1.0.0"
description = "WeKare Shared Infrastructure Components"
authors = ["WeKare Team <dev@wekare.com>"]
readme = "README.md"
packages = [
    {include = "config"},
    {include = "auth"},
    {include = "database"},
    {include = "app"},
    {include = "scripts"}
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
alembic = "^1.13.0"
boto3 = "^1.34.0"
jose = {extras = ["cryptography"], version = "^1.0.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"

[tool.poetry.group.test.dependencies]
httpx = "^0.25.2"
pytest-mock = "^3.12.0" 